requirements-cloudconvert.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.8 